requirements.txt
fastapi
uvicorn[standard]
jinja2
pydantic-settings
python-multipart
torch
transformers
ctransformers
llama-cpp-python
safetensors
gitpython
sqlmodel==0.0.21
aiosqlite==0.20.0
